Standardise styling for ChallengeCompletedDialog
