--- frontend/src/components/challengeCompletedModal/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ChallengeCompletedModal';
